Locate modules on the load-path
We have to do this for the libraries used by "ffi libraries"
as well as for our own "ffi-module.so".

Without doing it for the latter, we used to get:

  $ emacs -Q --batch -L . -f batch-byte-compile ffi.el

  In toplevel form:
  ffi.el:31:2: Error: Module could not be opened: \
  "ffi-module.so", "ffi-module.so: cannot open shared object file: \
  No such file or directory"

Note that in this case we are actually in the directory that contains
the module but never-the-less it doesn't work with a relative name.